backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
motor==3.3.0
pymongo==4.5.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.3.0
email-validator==2.0.0
python-multipart==0.0.6
httpx
